refactor: break up remote.ts into 9 focused modules
Extract 10 distinct responsibilities from the 3,134-line remote.ts into
separate modules with clean boundaries:

- remote-types.ts: shared interfaces and RelayContext type (~235 lines)
- remote-auth-source.ts: credential source detection (~37 lines)
- remote-provider-usage.ts: quota fetching/caching (~304 lines)
- remote-input.ts: attachment handling (~103 lines)
- remote-footer.ts: TUI footer rendering + pure layout utils (~170 lines)
- remote-heartbeat.ts: heartbeat construction/lifecycle (~97 lines)
- remote-exec-handler.ts: web exec command dispatch (~445 lines)
- remote-ask-user.ts: AskUserQuestion tool + dual TUI/web race (~376 lines)
- remote-plan-mode.ts: plan_mode tool + dual TUI/web race (~426 lines)

Orchestrator (remote.ts) reduced to 938 lines — handles RelayContext
creation, Socket.IO lifecycle, event forwarding, and module wiring.

Introduces RelayContext as an explicit injectable shared state object,
replacing implicit closure captures. All extracted modules receive it
as a parameter, enabling independent unit testing.

New tests: 32 tests across 3 files (footer utils, input parsing,
question sanitization). Zero behavioral changes — purely mechanical.

Design doc: docs/superpowers/specs/2026-03-15-remote-ts-refactor-design.md
